+++ src/main/java/com/project/tableforyou/domain/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.project.tableforyou.redis.domain;
+package com.project.tableforyou.domain.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
