+++ src/main/java/com/project/tableforyou/jwt/filter/JwtAuthenticationFilter.java
@@ -5,8 +5,8 @@ import com.project.tableforyou.config.auth.PrincipalDetails;
 import com.project.tableforyou.domain.dto.LoginDto;
 import com.project.tableforyou.handler.authFailureHandler.CustomAuthFailureHandler;
 import com.project.tableforyou.jwt.JwtUtil;
-import com.project.tableforyou.redis.domain.RefreshTokenDto;
-import com.project.tableforyou.redis.service.RefreshTokenService;
+import com.project.tableforyou.domain.dto.RefreshTokenDto;
+import com.project.tableforyou.service.RefreshTokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletInputStream;
